refactor: main 파일의 네비게이션 바 로직을 main_screen 파일로 분리
